chore: update utils + tests
