+++ server/utils/utils.test.ts
@@ -1,14 +1,4 @@
-import {
-  convertToTitleCase,
-  initialiseName,
-  formatDate,
-  formatDateTimeString,
-  formatDateOrDefault,
-  generateBasicAuthHeader,
-  getEstablishmentLinksData,
-} from './utils'
-
-import { DateFormats } from './enums'
+import { convertToTitleCase, generateBasicAuthHeader, getEstablishmentLinksData, initialiseName } from './utils'
 
 describe('convert to title case', () => {
   it.each([
@@ -65,46 +55,6 @@ describe('initialise name', () => {
     ['Double barrelled', 'Robert-John Smith-Jones-Wilson', 'R. Smith-Jones-Wilson'],
   ])('%s initialiseName(%s, %s)', (_: string, a: string, expected: string) => {
     expect(initialiseName(a)).toEqual(expected)
-  })
-})
-
-describe('Dates', () => {
-  let date: Date
-  let from: string
-  let to: string
-  let formattedFrom: string
-  let formattedTo: string
-
-  beforeEach(() => {
-    date = new Date('2025-12-01T10:30:00')
-    from = '2023-10-25T08:30:00'
-    to = '2023-10-25T11:45:00'
-    formattedFrom = '8.30am'
-    formattedTo = '11.45am'
-  })
-
-  afterEach(() => {
-    date = null
-    from = null
-    to = null
-    formattedFrom = null
-    formattedTo = null
-  })
-
-  it('it should return a string formatted version of the provided Date object', () => {
-    expect(formatDate(date, DateFormats.PRETTY_DATE)).toEqual('Monday 1 December, 2025')
-  })
-
-  it('it should return a string combining the provided from and to times in the expected format', () => {
-    expect(formatDateTimeString(from, to, DateFormats.PRETTY_TIME)).toEqual(`${formattedFrom} to ${formattedTo}`)
-  })
-
-  it('it should return a formatted time string in the expected format of x.xxam', () => {
-    expect(formatDateOrDefault('', DateFormats.PRETTY_TIME, from)).toEqual('8.30am')
-  })
-
-  it('it should return the provided placeHolder value when an invalid time format it provided', () => {
-    expect(formatDateOrDefault('', DateFormats.PRETTY_TIME, 'invalid')).toEqual('')
   })
 })
